+++ server/py/services/api/crud/runtimes/nuclio/function.py
@@ -249,14 +249,9 @@ def _compile_function_config(
 
     :return: function name, project name, nuclio function config
     """
-    if mlrun.mlconf.is_iguazio_v4_mode():
-        function = services.api.crud.secrets.Secrets.mount_secret_token_to_runtime(
-            function,
-            token_name=function.spec.auth.get("token_name"),
-            username=auth_info.username,
-        )
+    _enrich_config_spec(function, username=auth_info.username)
     # resolve env vars before compiling the nuclio spec, as we need to set them in the spec
-    env_dict, external_source_env_dict = _resolve_env_vars(function, auth_info)
+    env_dict, external_source_env_dict = _resolve_env_vars(function)
 
     project = function.metadata.project
     tag = function.metadata.tag
@@ -367,15 +362,20 @@ def _apply_escaped_config(config, parent_key, items: dict):
         mlrun.utils.update_in(config, f"{parent_key}.\\{key}\\", value)
 
 
-def _resolve_env_vars(function, auth_info=None):
+def _enrich_config_spec(function, username):
     # Add secret configurations to function's pod spec, if secret sources were added.
     # Needs to be here, since it adds env params, which are handled in the next lines.
     # This only needs to run if we're running within k8s context. If running in Docker, for example, skip.
     if framework.utils.singletons.k8s.get_k8s_helper(
         silent=True
     ).is_running_inside_kubernetes_cluster():
-        _add_secrets_config_to_function_spec(function)
+        token_name = function.spec.to_dict().get("auth", None).get("token_name", None)
+        _add_secrets_config_to_function_spec(
+            function, token_name=token_name, username=username
+        )
 
+
+def _resolve_env_vars(function):
     env_dict, external_source_env_dict = function._get_nuclio_config_spec_env()
     mlrun.auth.utils.enrich_auth_env(env_dict)
 
@@ -680,6 +680,8 @@ def _set_function_name(function, config, project, tag):
 
 def _add_secrets_config_to_function_spec(
     function: mlrun.runtimes.nuclio.function.RemoteRuntime,
+    token_name=None,
+    username=None,
 ):
     handler = services.api.runtime_handlers.BaseRuntimeHandler
     if function.kind in [
@@ -695,6 +697,8 @@ def _add_secrets_config_to_function_spec(
             function,
             project_name=function.metadata.project,
             encode_key_names=False,
+            token_name=token_name,
+            username=username,
         )
 
     elif function.kind == mlrun.runtimes.RuntimeKinds.serving:
@@ -715,10 +719,16 @@ def _add_secrets_config_to_function_spec(
                 function._secrets.get_k8s_secrets(),
                 function,
                 project_name=function.metadata.project,
+                token_name=token_name,
+                username=username,
             )
         else:
             handler.add_k8s_secrets_to_spec(
-                None, function, project_name=function.metadata.project
+                None,
+                function,
+                project_name=function.metadata.project,
+                token_name=token_name,
+                username=username,
             )
 
     else:
