+++ server/py/services/api/crud/secrets.py
@@ -814,61 +814,6 @@ class Secrets(
     def _generate_uuid() -> str:
         return str(uuid.uuid4())
 
-    @staticmethod
-    def mount_secret_token_to_runtime(
-        runtime: mlrun.runtimes.base.BaseRuntime, token_name: str, username: str
-    ):
-        # Validation that the secret exists is done in the ServerSideLauncher
-        secret = framework.utils.singletons.k8s.get_k8s_helper()._get_user_token_secret(
-            username=username, token_name=token_name
-        )
-
-        # Remove any existing auth secret volumes/mounts
-        services.api.crud.secrets.Secrets._remove_auth_secret_volumes(runtime)
-
-        # In case the secret was not found (which should not happen because of the prior validation), we do not mount it
-        if secret:
-            runtime.apply(
-                mlrun.mounts.mount_secret(
-                    secret.metadata.name,
-                    mount_path=mlrun.common.constants.MLRUN_JOB_AUTH_SECRET_PATH,
-                    items=[
-                        {
-                            "key": "tokensFile",
-                            "path": mlrun.common.constants.MLRUN_JOB_AUTH_SECRET_FILE,
-                        }
-                    ],
-                )
-            )
-        return runtime
-
-    @staticmethod
-    def _remove_auth_secret_volumes(runtime):
-        volumes = runtime.spec.volumes or []
-        mounts = runtime.spec.volume_mounts or []
-
-        volumes_to_remove = set()
-
-        # Identify volumes to remove
-        for vol in volumes:
-            if "secret" in vol:
-                secret = vol["secret"]
-                secret_name = secret.get("secretName", "")
-
-                # Pattern of auth secret volumes
-                if secret_name.startswith("mlrun-auth-secrets"):
-                    volumes_to_remove.add(vol["name"])
-
-        # Filter out only the matched volumes
-        runtime.spec.volumes = [
-            v for v in volumes if v["name"] not in volumes_to_remove
-        ]
-
-        # Filter out matching mounts
-        runtime.spec.volume_mounts = [
-            m for m in mounts if m["name"] not in volumes_to_remove
-        ]
-
 
 def get_project_secret_provider(project: str) -> typing.Callable:
     """Implement secret provider for handle the related project secret on the API side.
