+++ server/py/services/api/runtime_handlers/base.py
@@ -299,6 +299,8 @@ class BaseRuntimeHandler(ABC):
         self,
         runtime: mlrun.runtimes.pod.KubeResource,
         project_name: Optional[str] = None,
+        token_name: Optional[str] = None,
+        username: Optional[str] = None,
     ):
         if runtime._secrets:
             if runtime._secrets.has_vault_source():
@@ -313,9 +315,17 @@ class BaseRuntimeHandler(ABC):
                 runtime._secrets.get_k8s_secrets(),
                 runtime,
                 project_name=project_name,
+                token_name=token_name,
+                username=username,
             )
         else:
-            self.add_k8s_secrets_to_spec(None, runtime, project_name=project_name)
+            self.add_k8s_secrets_to_spec(
+                None,
+                runtime,
+                project_name=project_name,
+                token_name=token_name,
+                username=username,
+            )
 
     @staticmethod
     def add_vault_params_to_spec(
@@ -402,7 +412,12 @@ class BaseRuntimeHandler(ABC):
         runtime: mlrun.runtimes.pod.KubeResource,
         project_name: Optional[str] = None,
         encode_key_names: bool = True,
+        token_name: Optional[str] = None,
+        username: Optional[str] = None,
     ):
+        # In IG4, we add auth token secret as volumes and volumes mounts
+        BaseRuntimeHandler._mount_secret_token_to_runtime(runtime, token_name, username)
+
         # Check if we need to add the keys of a global secret. Global secrets are intentionally added before
         # project secrets, to allow project secret keys to override them
         global_secret_name = (
@@ -464,6 +479,32 @@ class BaseRuntimeHandler(ABC):
         # can be initialized with those env variables as secrets
         if not encode_key_names and secrets.keys():
             runtime.set_env("MLRUN_PROJECT_SECRETS_LIST", ",".join(secrets.keys()))
+
+    @staticmethod
+    def _mount_secret_token_to_runtime(
+        runtime: mlrun.runtimes.base.BaseRuntime, token_name: str, username: str
+    ):
+        # Validation that the secret exists is done in the ServerSideLauncher
+        secret = framework.utils.singletons.k8s.get_k8s_helper()._get_user_token_secret(
+            username=username, token_name=token_name
+        )
+
+        # In case the secret was not found (such as in IG3), we do not mount it
+        if secret:
+            # Remove any existing auth secret volumes/mounts
+            runtime.remove_auth_secret_volumes()
+            runtime.apply(
+                mlrun.mounts.mount_secret(
+                    secret.metadata.name,
+                    mount_path=mlrun.common.constants.MLRUN_JOB_AUTH_SECRET_PATH,
+                    items=[
+                        {
+                            "key": "tokensFile",
+                            "path": mlrun.common.constants.MLRUN_JOB_AUTH_SECRET_FILE,
+                        }
+                    ],
+                )
+            )
 
     @staticmethod
     def are_resources_coupled_to_run_object() -> bool:
