+++ tests/runtimes/test_base.py
@@ -326,3 +326,97 @@ class TestAutoMount:
         self._execute_run(runtime)
 
         rundb_mock.assert_runtime_categories(expected_categories)
+
+    @pytest.mark.parametrize(
+        "initial_volumes,initial_mounts,expected_volumes,expected_mounts",
+        [
+            # No volumes or mounts
+            ([], [], [], []),
+            # Only auth secret volume → removed
+            (
+                [
+                    {
+                        "name": "secret",
+                        "secret": {
+                            "secretName": "mlrun-auth-secrets-123",
+                            "items": [],
+                        },
+                    }
+                ],
+                [{"name": "secret", "mountPath": "/var/mlrun-secrets/auth"}],
+                [],
+                [],
+            ),
+            # Non-auth secret volume → preserved
+            (
+                [
+                    {
+                        "name": "user-secret",
+                        "secret": {
+                            "secretName": "my-user-secret",
+                            "items": [],
+                        },
+                    }
+                ],
+                [{"name": "user-secret", "mountPath": "/some/path"}],
+                [
+                    {
+                        "name": "user-secret",
+                        "secret": {
+                            "secretName": "my-user-secret",
+                            "items": [],
+                        },
+                    }
+                ],
+                [{"name": "user-secret", "mountPath": "/some/path"}],
+            ),
+            # Mixed auth + non-auth → remove only auth
+            (
+                [
+                    {
+                        "name": "auth-secret",
+                        "secret": {
+                            "secretName": "mlrun-auth-secrets-abc",
+                            "items": [],
+                        },
+                    },
+                    {
+                        "name": "other-secret",
+                        "secret": {
+                            "secretName": "other-secret",
+                            "items": [],
+                        },
+                    },
+                ],
+                [
+                    {"name": "auth-secret", "mountPath": "/var/mlrun-secrets/auth"},
+                    {"name": "other-secret", "mountPath": "/some/path"},
+                ],
+                [
+                    {
+                        "name": "other-secret",
+                        "secret": {
+                            "secretName": "other-secret",
+                            "items": [],
+                        },
+                    }
+                ],
+                [{"name": "other-secret", "mountPath": "/some/path"}],
+            ),
+        ],
+    )
+    def test_remove_auth_secret_volumes(
+        self,
+        initial_volumes,
+        initial_mounts,
+        expected_volumes,
+        expected_mounts,
+    ):
+        runtime = self._generate_runtime()
+        runtime.spec.volumes = initial_volumes.copy()
+        runtime.spec.volume_mounts = initial_mounts.copy()
+
+        runtime.remove_auth_secret_volumes()
+
+        assert runtime.spec.volumes == expected_volumes
+        assert runtime.spec.volume_mounts == expected_mounts
